took out posts next page load logic to hook
